+++ src/MPSegment.hpp
@@ -66,7 +66,7 @@ namespace CppJieba
                     return false;
                 }
                 string tmp;
-                for(uint i = 0; i < segWordInfos.size(); i++)
+                for(size_t i = 0; i < segWordInfos.size(); i++)
                 {
                     if(TransCode::encode(segWordInfos[i].word, tmp))
                     {
@@ -123,7 +123,7 @@ namespace CppJieba
                 for(Unicode::const_iterator it = begin; it != end; it++)
                 {
                     SegmentChar schar(*it);
-                    uint i = it - begin;
+                    size_t i = it - begin;
                     _trie.find(it, end, i, schar.dag);
                     //DagType::iterator dagIter;
                     if(schar.dag.end() ==  schar.dag.find(i))
@@ -148,7 +148,7 @@ namespace CppJieba
                     segContext[i].weight = MIN_DOUBLE;
                     for(DagType::const_iterator it = segContext[i].dag.begin(); it != segContext[i].dag.end(); it++)
                     {
-                        uint nextPos = it->first;
+                        size_t nextPos = it->first;
                         const TrieNodeInfo* p = it->second;
                         double val = 0.0;
                         if(nextPos + 1 < segContext.size())
@@ -176,7 +176,7 @@ namespace CppJieba
             }
             bool _cut(SegmentContext& segContext, vector<TrieNodeInfo>& res)const
             {
-                uint i = 0;
+                size_t i = 0;
                 while(i < segContext.size())
                 {
                     const TrieNodeInfo* p = segContext[i].pInfo;
